Extract scoring out of ImageGrid into useGridScoring
Scoring, guest scoring and the repositioning that follows a score change
move to their own composable: 640 lines out of ImageGrid, taking a deps
object in the same shape useGridFetch and useStackOrdering already use.

The repositioning is the part worth keeping together with the scoring.
Setting a score under a score or smart-score sort leaves the picture in
the wrong row, so these functions splice it to its new position instead
of refetching, which is what stops a rating pass flickering.

useStackOrdering is created before this one and needs its range
invalidation, so that one reference is deferred rather than passed by
value.
